Add review optimization and refactoring
